+++ src/net/fexcraft/app/fmt/FMTB.java
@@ -33,18 +33,18 @@ import net.fexcraft.app.fmt.porters.PorterManager;
 import net.fexcraft.app.fmt.ui.Dialog;
 import net.fexcraft.app.fmt.ui.FontRenderer;
 import net.fexcraft.app.fmt.ui.UserInterface;
-import net.fexcraft.app.fmt.ui.general.NFC;
-import net.fexcraft.app.fmt.ui.re.Bottombar;
-import net.fexcraft.app.fmt.ui.re.ControlsAdjuster;
-import net.fexcraft.app.fmt.ui.re.DialogBox;
-import net.fexcraft.app.fmt.ui.re.SettingsBox;
-import net.fexcraft.app.fmt.ui.re.TextField;
-import net.fexcraft.app.fmt.ui.re.Toolbar;
-import net.fexcraft.app.fmt.ui.re.editor.Editor;
-import net.fexcraft.app.fmt.ui.re.editor.GeneralEditor;
-import net.fexcraft.app.fmt.ui.re.editor.ModelGroupEditor;
-import net.fexcraft.app.fmt.ui.re.editor.PreviewEditor;
-import net.fexcraft.app.fmt.ui.re.editor.TextureEditor;
+import net.fexcraft.app.fmt.ui.editor.Editor;
+import net.fexcraft.app.fmt.ui.editor.GeneralEditor;
+import net.fexcraft.app.fmt.ui.editor.ModelGroupEditor;
+import net.fexcraft.app.fmt.ui.editor.PreviewEditor;
+import net.fexcraft.app.fmt.ui.editor.TextureEditor;
+import net.fexcraft.app.fmt.ui.general.Bottombar;
+import net.fexcraft.app.fmt.ui.general.ControlsAdjuster;
+import net.fexcraft.app.fmt.ui.general.DialogBox;
+import net.fexcraft.app.fmt.ui.general.SettingsBox;
+import net.fexcraft.app.fmt.ui.general.TextField;
+import net.fexcraft.app.fmt.ui.general.Toolbar;
+import net.fexcraft.app.fmt.ui.old.NFC;
 import net.fexcraft.app.fmt.utils.Backups;
 import net.fexcraft.app.fmt.utils.DiscordUtil;
 import net.fexcraft.app.fmt.utils.GGR;
