+++ src/net/fexcraft/app/fmt/ui/general/ControlsAdjuster.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package net.fexcraft.app.fmt.ui.re;
+package net.fexcraft.app.fmt.ui.general;
 
 import org.lwjgl.input.Keyboard;
 import org.lwjgl.opengl.GL11;
