+++ src/net/fexcraft/app/fmt/ui/general/HoverMenu.java
@@ -2,48 +2,37 @@ package net.fexcraft.app.fmt.ui.general;
 
 import java.util.ArrayList;
 
-import net.fexcraft.app.fmt.ui.Element;
+import net.fexcraft.app.fmt.ui.NewElement;
 
-/**
- * 
- * @author Ferdinand Calo' (FEX___96)
- *
- */
-public abstract class HoverMenu extends Element {
+public abstract class HoverMenu extends NewElement {
 	
-	public static final ArrayList<HoverMenu> arrlist = new ArrayList<>();
+	public static final ArrayList<HoverMenu> MENUS = new ArrayList<>();
 
-	public HoverMenu(Element root, String id, int width){
-		super(root, id); this.setSize(width, 2).setLevel(root.getLevel() + 1);
-		this.setPosition(root.x - 2, root.y + root.height);
-		this.setTexPosSize("ui/background_dark", 0, 0, 64, 64);
-		this.addButtons(); HoverMenu.arrlist.add(this);
+	public HoverMenu(NewElement root, String id, int width){
+		super(root, id, "hovermenu"); MENUS.add(this); this.setSize(width, 0).setColor(0xffc7c7c7).setVisible(false); this.addButtons();
 	}
-
+	
 	public abstract void addButtons();
-
+	
+	@Override
+	public NewElement setVisible(boolean bool){
+		for(NewElement elm : elements) elm.setVisible(bool); this.visible = bool; return this;
+	}
+	
+	public static boolean anyMenuHovered(){
+		for(HoverMenu menu : MENUS) if(menu.hovered && menu.visible) return true; return false;
+	}
+	
 	@Override
 	public void renderSelf(int rw, int rh){
-		if(!hovered && !root.isHovered()){
-			this.setVisible(false);
-		}
-		this.height = 2;
-		for(Element elm : elements){ height += elm.height + 2; }
+		if(!hovered && !root.isHovered()){ this.setVisible(false); }
+		for(NewElement elm : elements){ height += elm.height + 2; }
 		this.renderSelfQuad();
 	}
 	
 	@Override
-	public Element setVisible(boolean bool){
-		for(Element elm : elements) elm.setVisible(bool); this.visible = bool; return this;
-	}
-	
-	@Override
-	protected void realignToRoot(int index){
-		this.setPosition(root.x - 2, root.y + root.height);
-	}
-	
-	public static boolean anyMenuHovered(){
-		for(HoverMenu menu : arrlist) if(menu.hovered && menu.visible) return true; return false;
+	public NewElement repos(){
+		x = root.x + xrel; y = root.y + yrel + root.height; clearVertexes(); height = 0; for(NewElement elm : elements) elm.repos(); return this;
 	}
 
 }
