+++ src/net/fexcraft/app/fmt/ui/general/Toolbar.java
@@ -5,12 +5,11 @@ import java.io.File;
 
 import net.fexcraft.app.fmt.FMTB;
 import net.fexcraft.app.fmt.porters.PorterManager;
-import net.fexcraft.app.fmt.ui.Element;
+import net.fexcraft.app.fmt.ui.NewElement;
 import net.fexcraft.app.fmt.ui.UserInterface;
-import net.fexcraft.app.fmt.ui.general.NFC.AfterTask;
-import net.fexcraft.app.fmt.ui.general.NFC.ChooserMode;
-import net.fexcraft.app.fmt.ui.re.DialogBox;
-import net.fexcraft.app.fmt.ui.re.editor.Editor;
+import net.fexcraft.app.fmt.ui.editor.Editor;
+import net.fexcraft.app.fmt.ui.old.NFC.AfterTask;
+import net.fexcraft.app.fmt.ui.old.NFC.ChooserMode;
 import net.fexcraft.app.fmt.ui.tree.RightTree;
 import net.fexcraft.app.fmt.utils.GGR;
 import net.fexcraft.app.fmt.utils.HelperCollector;
@@ -29,71 +28,96 @@ import net.fexcraft.app.fmt.wrappers.ShapeboxWrapper;
 import net.fexcraft.app.fmt.wrappers.TexrectWrapperA;
 import net.fexcraft.app.fmt.wrappers.TexrectWrapperB;
 import net.fexcraft.app.fmt.wrappers.TurboList;
+import net.fexcraft.lib.common.math.RGB;
 import net.fexcraft.lib.common.utils.Print;
 
-/**
- * 
- * @author Ferdinand Calo' (FEX___96)
- *
- */
-public class Toolbar extends Element {
+public class Toolbar extends NewElement {
 
-	private int oldwidth;
-	
 	public Toolbar(){
-		super(null, "toolbar");
-		TextureManager.loadTexture("icons/toolbar/info", null);
-		TextureManager.loadTexture("icons/toolbar/new", null);
-		TextureManager.loadTexture("icons/toolbar/open", null);
-		TextureManager.loadTexture("icons/toolbar/save", null);
-		TextureManager.loadTexture("icons/toolbar/profile", null);
-		TextureManager.loadTexture("icons/toolbar/settings", null);
-		this.setPosition(0, 60).setLevel(-20).setTexPosSize("ui/background_dark", 0, 0, 32, 32).setSize(100, 30);
+		super(null, "toolbar", "toolbar"); hovercolor = RGB.WHITE;
+		this.setPosition(0, 0, -20).setSize(100, 30).setColor(0xff484848).setBorder(0xff32a852, 0xffeb4034, 3, false, true);
+		//
+		this.elements.add(new Icon(this, "info", "toolbar:icon", "icons/toolbar/info", 32, 0, 0){
+			@Override
+			protected boolean processButtonClick(int x, int y, boolean left){
+				DialogBox.notAvailableYet(); return true;
+			}
+		});
+		this.elements.add(new Icon(this, "new_file", "toolbar:icon", "icons/toolbar/new", 32, 0, 0){
+			@Override
+			protected boolean processButtonClick(int x, int y, boolean left){
+				SaveLoad.openNewModel(); return true;
+			}
+		});
+		this.elements.add(new Icon(this, "open_file", "toolbar:icon", "icons/toolbar/open", 32, 0, 0){
+			@Override
+			protected boolean processButtonClick(int x, int y, boolean left){
+				SaveLoad.openModel(); return true;
+			}
+		});
+		this.elements.add(new Icon(this, "save_file", "toolbar:icon", "icons/toolbar/save", 32, 0, 0){
+			@Override
+			protected boolean processButtonClick(int x, int y, boolean left){
+				SaveLoad.saveModel(false, false); return true;
+			}
+		});
+		this.elements.add(new Icon(this, "profile", "toolbar:icon", "icons/toolbar/profile", 32, 0, 0){
+			@Override
+			protected boolean processButtonClick(int x, int y, boolean left){
+				DialogBox.notAvailableYet(); return true;
+			}
+		});
+		this.elements.add(new Icon(this, "settings", "toolbar:icon", "icons/toolbar/settings", 32, 0, 0){
+			@Override
+			protected boolean processButtonClick(int x, int y, boolean left){
+				UserInterface.SETTINGSBOX.show(); return true;
+			}
+		});
 		//
 		int btsz = 96, bthg = 28;
-		this.elements.add(new Button(this, "file", btsz, bthg, 0, 0){
+		this.elements.add(new Button(this, "file", null, btsz, bthg, 0, 0){
 			@Override
 			public void setupSubmenu(){
 				HoverMenu menu = new HoverMenu(this, "menu", 100){
 					@Override
 					public void addButtons(){
-						this.elements.add(new Button(this, "new_model", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "new_model", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								SaveLoad.openNewModel(); return true;
 							}
 						}.setText("New Model", false));
-						this.elements.add(new Button(this, "open", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "open", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								SaveLoad.openModel(); return true;
 							}
 						}.setText("Open Model", false));
-						this.elements.add(new Button(this, "save", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "save", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								SaveLoad.saveModel(false, false); return true;
 							}
 						}.setText("Save Model", false));
-						this.elements.add(new Button(this, "save_as", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "save_as", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								SaveLoad.saveModel(true, false); return true;
 							}
 						}.setText("Save as...", false));
-						this.elements.add(new Button(this, "import", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "import", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								PorterManager.handleImport(); return true;
 							}
 						}.setText("Import <<", false));
-						this.elements.add(new Button(this, "export", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "export", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								PorterManager.handleExport(); return true;
 							}
 						}.setText("Export >>", false));
-						this.elements.add(new Button(this, "edit", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "exit", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								SaveLoad.checkIfShouldSave(true, false); return true;
@@ -104,44 +128,44 @@ public class Toolbar extends Element {
 				this.elements.add(menu);
 			}
 		}.setText("File", true));
-		this.elements.add(new Button(this, "utils", btsz, bthg, 0, 0){
+		this.elements.add(new Button(this, "utils", null, btsz, bthg, 0, 0){
 			@Override
 			public void setupSubmenu(){
 				HoverMenu menu = new HoverMenu(this, "menu", 100){
 					@Override
 					public void addButtons(){
-						this.elements.add(new Button(this, "undo", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "undo", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								DialogBox.notAvailableYet(); return true;
 							}
 						}.setText("Undo", false));
-						this.elements.add(new Button(this, "redo", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "redo", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								DialogBox.notAvailableYet(); return true;
 							}
 						}.setText("Redo", false));
-						this.elements.add(new Button(this, "create_gif", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "create_gif", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								ImageHelper.createGif(false); return true;
 							}
 						}.setText("Create Gif", false));
-						this.elements.add(new Button(this, "screenshot", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "screenshot", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								ImageHelper.takeScreenshot(true); return true;
 							}
 						}.setText("Screenshot", false));
-						this.elements.add(new Button(this, "reset", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "reset", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.ggr = new GGR(FMTB.get(), 0, 4, 4); FMTB.ggr.rotation.xCoord = 45;
 								return true;
 							}
 						}.setText("Reset Camera", false));
-						this.elements.add(new Button(this, "calc_size", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "calc_size", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								long l = 0, ll = 0;
@@ -164,13 +188,13 @@ public class Toolbar extends Element {
 				this.elements.add(menu);
 			}
 		}.setText("Utils", true));
-		this.elements.add(new Button(this, "editor", btsz, bthg, 0, 0){
+		this.elements.add(new Button(this, "editor", null, btsz, bthg, 0, 0){
 			@Override
 			public void setupSubmenu(){
 				HoverMenu menu = new HoverMenu(this, "menu", 100){
 					@Override
 					public void addButtons(){
-						this.elements.add(new Button(this, "copy_selection", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "copy_selection", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.MODEL.copyAndSelect(); return true;
@@ -178,17 +202,11 @@ public class Toolbar extends Element {
 						}.setText("Copy Selected", false));
 						for(int i = 0; i < 3; i++){
 							String str = i == 0 ? "x" : i == 1 ? "y" : "z"; int j = i; String[] arr = new String[]{ "[L/R]", "[U/D]", "[F/B]"};
-							this.elements.add(new Button(this, "flip_" + str, 20, 26, 0, 0){
+							this.elements.add(new Button(this, "flip_" + str, null, 20, 26, 0, 0){
 								@Override protected boolean processButtonClick(int x, int y, boolean left){ FMTB.MODEL.flipShapeboxes(j); return true; }
 							}.setText("Flip (" + str.toUpperCase() + ") " + arr[j], false));
 						}
-						this.elements.add(new Button(this, "hr0", 20, 4, 0, 0){
-							@Override
-							protected boolean processButtonClick(int x, int y, boolean left){
-								/** Generic Placeholder Space/<HR> **/return true;
-							}
-						}.setText("", false));
-						this.elements.add(new Button(this, "controls", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "controls", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								UserInterface.CONTROLS.show(); return true;
@@ -199,32 +217,31 @@ public class Toolbar extends Element {
 				this.elements.add(menu);
 			}
 		}.setText("Editor", true));
-		this.elements.add(new Button(this, "shapeditor", btsz, bthg, 0, 0){
+		this.elements.add(new Button(this, "shapeditor", null, btsz, bthg, 0, 0){
 			@Override
 			public void setupSubmenu(){
 				HoverMenu menu = new HoverMenu(this, "menu", 100){
 					@Override
 					public void addButtons(){
-						this.elements.add(new Button(this, "toggle", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "toggle", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								Editor.toggleAll(); return true;
 							}
-						}.setText("Toggle Visibility", false));
-						this.elements.add(new Button(this, "hr0", 20, 4, 0, 0).setText("", false));
-						this.elements.add(new Button(this, "general", 20, 26, 0, 0){
+						}.setText("Hide Editors", false));
+						this.elements.add(new Button(this, "general", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								Editor.show("general_editor"); return true;
 							}
 						}.setText("General Editor", false));
-						this.elements.add(new Button(this, "texture", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "texture", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								Editor.show("texture_editor"); return true;
 							}
 						}.setText("Texture Editor (beta)", false));
-						this.elements.add(new Button(this, "model_group", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "model_group", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								Editor.show("model_group_editor"); return true;
@@ -235,7 +252,7 @@ public class Toolbar extends Element {
 				this.elements.add(menu);
 			}
 		}.setText("Shapeditor", true));
-		this.elements.add(new Button(this, "shapelist", btsz, bthg, 0, 0){
+		this.elements.add(new Button(this, "shapelist", null, btsz, bthg, 0, 0){
 			@Override
 			protected boolean processButtonClick(int x, int y, boolean left){
 				RightTree.toggle("modeltree"); return true;
@@ -245,48 +262,42 @@ public class Toolbar extends Element {
 				HoverMenu menu = new HoverMenu(this, "menu", 100){
 					@Override
 					public void addButtons(){
-						this.elements.add(new Button(this, "add_box", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "add_box", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.MODEL.add(new BoxWrapper(FMTB.MODEL), null, true);
 								this.root.setVisible(false); return true;
 							}
 						}.setText("Add Generic Box", false));
-						this.elements.add(new Button(this, "add_shapebox", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "add_shapebox", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.MODEL.add(new ShapeboxWrapper(FMTB.MODEL), null, true);
 								this.root.setVisible(false); return true;
 							}
 						}.setText("Add Shapebox", false));
-						this.elements.add(new Button(this, "add_texrectb", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "add_texrectb", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.MODEL.add(new TexrectWrapperB(FMTB.MODEL), null, true);
 								this.root.setVisible(false); return true;
 							}
 						}.setText("Add TexRect [Basic]", false));
-						this.elements.add(new Button(this, "add_texrecta", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "add_texrecta", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.MODEL.add(new TexrectWrapperA(FMTB.MODEL), null, true);
 								this.root.setVisible(false); return true;
 							}
 						}.setText("Add TexRect [Adv.]", false));
-						this.elements.add(new Button(this, "add_cylinder", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "add_cylinder", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.MODEL.add(new CylinderWrapper(FMTB.MODEL), null, true);
 								this.root.setVisible(false); return true;
 							}
 						}.setText("Add Cylinder", false));
-						this.elements.add(new Button(this, "hr0", 20, 4, 0, 0){
-							@Override
-							protected boolean processButtonClick(int x, int y, boolean left){
-								/** Generic "Space/<hr>" Button. **/return true;
-							}
-						}.setText("", true));
-						this.elements.add(new Button(this, "add_group", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "add_group", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								String string = "group" + FMTB.MODEL.getCompound().size();
@@ -301,14 +312,14 @@ public class Toolbar extends Element {
 								this.root.setVisible(false); return true;
 							}
 						}.setText("Add Group", false));
-						this.elements.add(new Button(this, "add_marker", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "add_marker", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.MODEL.add(new MarkerWrapper(FMTB.MODEL), "markers", true);
 								this.root.setVisible(false); return true;
 							}
 						}.setText("Add Marker", false));
-						this.elements.add(new Button(this, "add_collisiongrid", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "add_collisiongrid", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.MODEL.add(new CollisionGridWrapper(FMTB.MODEL), null, true);
@@ -320,13 +331,13 @@ public class Toolbar extends Element {
 				this.elements.add(menu);
 			}
 		}.setText("Shapelist", true));
-		this.elements.add(new Button(this, "textures", btsz, bthg, 0, 0){
+		this.elements.add(new Button(this, "textures", null, btsz, bthg, 0, 0){
 			@Override
 			public void setupSubmenu(){
 				HoverMenu menu = new HoverMenu(this, "menu", 100){
 					@Override
 					public void addButtons(){
-						this.elements.add(new Button(this, "select", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "select", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								UserInterface.FILECHOOSER.show(new String[]{ "Select a texture file." }, new File("./resources/textures"), new AfterTask(){
@@ -341,7 +352,7 @@ public class Toolbar extends Element {
 								}, ChooserMode.PNG); return true;
 							}
 						}.setText("Select Texture", false));
-						this.elements.add(new Button(this, "edit", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "edit", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								if(FMTB.MODEL.texture == null) return true;
@@ -356,7 +367,7 @@ public class Toolbar extends Element {
 								} catch(Exception e){ e.printStackTrace(); } return true;
 							}
 						}.setText("Edit (External)", false));
-						this.elements.add(new Button(this, "remove", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "remove", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								if(FMTB.MODEL.texture != null && TextureManager.getTexture(FMTB.MODEL.texture, true) != null){
@@ -364,13 +375,7 @@ public class Toolbar extends Element {
 								} return true;
 							}
 						}.setText("Remove/Unload", false));
-						this.elements.add(new Button(this, "hr0", 20, 4, 0, 0){
-							@Override
-							protected boolean processButtonClick(int x, int y, boolean left){
-								/** Generic Space/<HR> Button **/ return true;
-							}
-						}.setText("", false));
-						this.elements.add(new Button(this, "texreset", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "texreset", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								FMTB.MODEL.getCompound().values().forEach(list -> list.forEach(turbo -> {
@@ -379,13 +384,13 @@ public class Toolbar extends Element {
 								FMTB.showDialogbox("Texture Positions Reset.", "ok",  null, DialogBox.NOTHING, null); return true;
 							}
 						}.setText("Reset TexPos.", false));
-						this.elements.add(new Button(this, "autopos", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "autopos", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								TextureUpdate.tryAutoPos(null); return true;
 							}
 						}.setText("Auto Position", false));
-						this.elements.add(new Button(this, "generate", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "generate", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								String texname = "temp/" + FMTB.MODEL.name;
@@ -403,7 +408,7 @@ public class Toolbar extends Element {
 				this.elements.add(menu);
 			}
 		}.setText("Texture", true));
-		this.elements.add(new Button(this, "helpers", btsz, bthg, 0, 0){
+		this.elements.add(new Button(this, "helpers", null, btsz, bthg, 0, 0){
 			@Override
 			protected boolean processButtonClick(int x, int y, boolean left){
 				RightTree.toggle("helpertree"); return true;
@@ -413,19 +418,13 @@ public class Toolbar extends Element {
 				HoverMenu menu = new HoverMenu(this, "menu", 100){
 					@Override
 					public void addButtons(){
-						this.elements.add(new Button(this, "view", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "view", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								RightTree.show("helpertree"); return true;
 							}
 						}.setText("View Loaded", false));
-						this.elements.add(new Button(this, "hr0", 20, 4, 0, 0){
-							@Override
-							protected boolean processButtonClick(int x, int y, boolean left){
-								/** Generic Placeholder Space/<HR> **/return true;
-							}
-						}.setText("", false));
-						this.elements.add(new Button(this, "fmtb", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "fmtb", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								UserInterface.FILECHOOSER.show(new String[]{ "Select a Preview/Helper file." }, new File("./helpers"), new AfterTask(){
@@ -433,7 +432,7 @@ public class Toolbar extends Element {
 								}, ChooserMode.SAVEFILE_LOAD); return true;
 							}
 						}.setText("Load FMTB", false));
-						this.elements.add(new Button(this, "frame", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "frame", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								UserInterface.FILECHOOSER.show(new String[]{ "Select an Image file." }, new File("./helpers"), new AfterTask(){
@@ -441,7 +440,7 @@ public class Toolbar extends Element {
 								}, ChooserMode.HELPFRAMEIMG); return true;
 							}
 						}.setText("Load Frame", false));
-						this.elements.add(new Button(this, "import", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "import", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								UserInterface.FILECHOOSER.show(new String[]{ "Select a Preview/Helper file." }, new File("./helpers"), new AfterTask(){
@@ -449,13 +448,7 @@ public class Toolbar extends Element {
 								}, ChooserMode.IMPORT); return true;
 							}
 						}.setText("Load Imported", false));
-						this.elements.add(new Button(this, "hr1", 20, 4, 0, 0){
-							@Override
-							protected boolean processButtonClick(int x, int y, boolean left){
-								/** Generic Placeholder Space/<HR> **/return true;
-							}
-						}.setText("", false));
-						this.elements.add(new Button(this, "clear", 20, 26, 0, 0){
+						this.elements.add(new Button(this, "clear", null, 20, 26, 0, 0){
 							@Override
 							protected boolean processButtonClick(int x, int y, boolean left){
 								HelperCollector.LOADED.clear(); return true;
@@ -466,35 +459,28 @@ public class Toolbar extends Element {
 				this.elements.add(menu);
 			}
 		}.setText("Helpers", true));
-		this.elements.add(new Button(this, "exit", btsz, bthg, 0, 0){
+		this.elements.add(new Button(this, "exit", null, btsz, bthg, 0, 0){
 			@Override
 			protected boolean processButtonClick(int x, int y, boolean left){
 				SaveLoad.checkIfShouldSave(true, false); return true;
 			}
 		}.setText("Exit", true));
 		//
-		this.reorderElements();
-	}
-	
-	@SuppressWarnings("unused")
-	private void reorderElements(){
-		int start = 4, high = 3 + 60;
-		for(Element elm : elements){
-			if(start + elm.width >= width - 4){ start = 4; high += 34; }
-			elm.setPosition(start, /*elm instanceof Icon*/ false ? high - 2 : high); start += elm.width + 2;
-		}
-		this.height = high + 31;
+		realign();
 	}
 
-	@Override
-	public void renderSelf(int root_width, int root_height){
-		this.width = root_width; this.renderSelfQuad();
-		if(oldwidth != width){ oldwidth = width; this.reorderElements(); }
+	private void realign(){
+		int start = 4, high = 3;
+		for(NewElement elm : elements){
+			if(start + elm.width >= width - 4){ start = 4; high += 34; }
+			elm.setPosition(start, elm instanceof Icon ? high - 2 : high, null); start += elm.width + 2;
+		}
+		this.height = high + 31; for(NewElement elm : elements) elm.repos();
 	}
 	
 	@Override
-	protected boolean processButtonClick(int x, int y, boolean left){
-		return true;
+	public NewElement repos(){
+		width = UserInterface.width; x = xrel; y = yrel; realign(); return this.clearVertexes();
 	}
 
 }
